+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # docs
 
+## 0.0.3
+
+### Patch Changes
+
+- negative-space@2.5.4
+
 ## 0.0.2
 
 ### Patch Changes